move to domain model
