+++ dsp-meta/src/api/convert/mod.rs
@@ -1,4 +1,3 @@
 pub(crate) mod axum;
 pub mod hcl;
 pub(crate) mod rdf;
-pub mod serde;
