+++ dsp-meta/src/domain/mod.rs
@@ -1,1 +1,2 @@
+pub mod model;
 pub mod service;
