+++ dsp-meta/src/domain/model/draft_model.rs
